+++ core/lib/spree/core/engine.rb
@@ -69,6 +69,7 @@ module Spree
           ].each { |event_name| Spree::Bus.register(event_name) }
 
           Spree::OrderMailerSubscriber.new.subscribe_to(Spree::Bus)
+          Spree::ReimbursementMailerSubscriber.new.subscribe_to(Spree::Bus)
         end
       end
